Wrote 'blender-send-buffer' based on Marc's function 'blender-execute-file'.
doc:

Send the current buffer to the inferior Python process.

When the buffer is bound to a file and the current state of the
buffer has been saved, the corresponding file itself is loaded
using 'imp.load_source'; when the buffer is not bound to a file
or has been modified since it has been saved last, the buffer
content is sent.